smash or pass? LFT-TOP_CNT done!

+++ skopchenko.alexander/P4/topclock.h
@@ -6,7 +6,7 @@
 namespace skopchenko
 {
 
-  void topClock(int *matrix, size_t rows, size_t cols, size_t value, int top, size_t bottom, int left, size_t right);
+  void topClock(int *matrix, size_t rows, size_t cols, size_t value, size_t top, size_t bottom, size_t left, size_t right);
 
 }
